requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
jinja2==3.1.4
sentence-transformers==3.3.1
scikit-learn==1.5.2
numpy==2.0.2
pandas==2.2.3
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pyperclip==1.9.0
supabase==2.10.0
psycopg2-binary==2.9.10